requirements.txt
fastapi
uvicorn[standard]
gspread
pandas
python-dotenv
streamlit
groq
streamlit-autorefresh